feat: 域名服务接入

+++ src/main/java/com/tencent/msdk/dns/core/cache/Cache.java
@@ -2,6 +2,7 @@ package com.tencent.msdk.dns.core.cache;
 
 import android.text.TextUtils;
 
+import com.tencent.msdk.dns.BuildConfig;
 import com.tencent.msdk.dns.DnsService;
 import com.tencent.msdk.dns.base.log.DnsLog;
 import com.tencent.msdk.dns.core.Const;
@@ -22,22 +23,43 @@ public final class Cache implements ICache {
 
     private static final LookupCacheDao lookupCacheDao = LookupCacheDatabase.getInstance(DnsService.getAppContext()).lookupCacheDao();
 
-    private boolean getCachedIpEnable() {
+    private static boolean getCachedIpEnable() {
         return DnsService.getDnsConfig().cachedIpEnable;
     }
 
-    public static void readFromDb() {
-        List<LookupCache> allCache = lookupCacheDao.getAll();
-        ArrayList<LookupCache> expired = new ArrayList<LookupCache>();
-        for (LookupCache lookupCache : allCache) {
-            mHostnameIpsMap.put(lookupCache.hostname, lookupCache.lookupResult);
+    private static boolean isIntlServerDomain(String hostname) {
+        return BuildConfig.FLAVOR.equals("intl") && hostname.equals(BuildConfig.INIT_SERVERS_DOMAINS[0]);
+    }
 
-            if (lookupCache.isExpired()) {
-                expired.add(lookupCache);
+    public static void readFromDb() {
+        if (getCachedIpEnable()) {
+            List<LookupCache> allCache = lookupCacheDao.getAll();
+            ArrayList<LookupCache> expired = new ArrayList<LookupCache>();
+            for (LookupCache lookupCache : allCache) {
+                mHostnameIpsMap.put(lookupCache.hostname, lookupCache.lookupResult);
+
+                if (lookupCache.isExpired()) {
+                    expired.add(lookupCache);
+                }
+            }
+            // 内存读取后，清空本地已过期的缓存
+            lookupCacheDao.deleteLookupCaches(expired);
+        } else {
+            // 服务域名缓存读取
+            readServerFromDb();
+        }
+    }
+
+    private static void readServerFromDb() {
+        // 国际站域名服务按州分布，持久化缓存
+        if (BuildConfig.FLAVOR.equals("intl")) {
+            LookupCache serverCache = lookupCacheDao.get(BuildConfig.INIT_SERVERS_DOMAINS[0]);
+            mHostnameIpsMap.put(serverCache.hostname, serverCache.lookupResult);
+
+            if (serverCache.isExpired()) {
+                lookupCacheDao.delete(serverCache.hostname);
             }
         }
-        // 内存读取后，清空本地已过期的缓存
-        lookupCacheDao.deleteLookupCaches(expired);
     }
 
     @Override
@@ -60,7 +82,7 @@ public final class Cache implements ICache {
 
         DnsLog.d("Cache %s for %s", lookupResult, hostname);
         mHostnameIpsMap.put(hostname, lookupResult);
-        if (getCachedIpEnable()) {
+        if (getCachedIpEnable() || isIntlServerDomain(hostname)) {
             lookupCacheDao.insertLookupCache(new LookupCache(hostname, lookupResult));
         }
     }
